Refactoring the wikipedia section of the application to ensure that the parsing of a wikipedia response isnt done in the middle of an activity.
