+++ BluePlaquesLondon/app/src/main/java/com/upwardsnorthwards/blueplaqueslondon/activities/WikipediaActivity.java
@@ -30,37 +30,27 @@ package com.upwardsnorthwards.blueplaqueslondon.activities;
 
 import android.app.Activity;
 import android.content.Intent;
-import android.os.AsyncTask;
+import android.net.Uri;
 import android.os.Bundle;
+import android.view.KeyEvent;
 import android.view.Window;
 import android.webkit.WebChromeClient;
+import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 
 import com.upwardsnorthwards.blueplaqueslondon.BluePlaquesLondonApplication;
 import com.upwardsnorthwards.blueplaqueslondon.R;
+import com.upwardsnorthwards.blueplaqueslondon.model.IWikipediaModelDelegate;
 import com.upwardsnorthwards.blueplaqueslondon.model.Placemark;
+import com.upwardsnorthwards.blueplaqueslondon.model.WikipediaModel;
 import com.upwardsnorthwards.blueplaqueslondon.utils.BluePlaquesConstants;
 
-import org.apache.http.HttpResponse;
-import org.apache.http.HttpStatus;
-import org.apache.http.StatusLine;
-import org.apache.http.client.ClientProtocolException;
-import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.DefaultHttpClient;
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.net.URLEncoder;
-
-public class WikipediaActivity extends BaseActivity {
+public class WikipediaActivity extends BaseActivity implements IWikipediaModelDelegate {
 
     private WebView webView;
     private Placemark placemark;
+    private WikipediaModel wikipediaModel;
 
     @Override
     protected void onCreate(final Bundle savedInstanceState) {
@@ -74,17 +64,33 @@ public class WikipediaActivity extends BaseActivity {
             placemark = (Placemark) intent
                     .getParcelableExtra(BluePlaquesConstants.WIKIPEDIA_CLICKED_PARCLEABLE_KEY);
         }
-        new WikipediaModel().execute(placemark.getName(),
-                getString(R.string.wikipedia_url));
     }
 
     @Override
     protected void onResume() {
         super.onResume();
         setCustomTitleBarText(placemark.getTrimmedName());
+        wikipediaModel = new WikipediaModel();
+        wikipediaModel.setDelegate(this);
+        wikipediaModel.execute(placemark.getName(), getString(R.string.wikipedia_url));
     }
 
-    protected void onRetriveWikipediaUrlSuccess(final String url) {
+    @Override
+    protected void onPause() {
+        super.onPause();
+        wikipediaModel.onPause();
+    }
+
+    @Override
+    public boolean onKeyDown(int keyCode, KeyEvent event) {
+        if ((keyCode == KeyEvent.KEYCODE_BACK) && webView.canGoBack()) {
+            webView.goBack();
+            return true;
+        }
+        return super.onKeyDown(keyCode, event);
+    }
+
+    public void onRetriveWikipediaUrlSuccess(final String url) {
         final Activity activity = this;
         webView.setWebChromeClient(new WebChromeClient() {
             public void onProgressChanged(final WebView view, final int progress) {
@@ -100,73 +106,26 @@ public class WikipediaActivity extends BaseActivity {
         webView.loadUrl(url);
     }
 
-    protected void onRetriveWikipediaUrlFailure() {
+    public void onRetriveWikipediaUrlFailure() {
         final BluePlaquesLondonApplication app = (BluePlaquesLondonApplication) getApplication();
         app.trackEvent(BluePlaquesConstants.ERROR_CATEGORY,
                 BluePlaquesConstants.WIKIPEDIA_PAGE_LOAD_ERROR_EVENT,
                 placemark.getName());
     }
 
-    public class WikipediaModel extends AsyncTask<String, String, String> {
-
-        private static final String WIKIPEDIA_SEARCH_URL_FORMAT = "http://en.wikipedia.org/w/api.php?action=query&list=search&srsearch=%s&srprop=timestamp&format=json";
-        private static final String WIKIPEDIA_MODEL_ENCODING = "UTF-8";
-        private String responseUrl;
-
-        @Override
-        protected String doInBackground(final String... params) {
-            final String name = params[0];
-            responseUrl = params[1];
-            final HttpClient httpclient = new DefaultHttpClient();
-            String responseString = null;
-            try {
-                final String url = String.format(WIKIPEDIA_SEARCH_URL_FORMAT,
-                        URLEncoder.encode(name, WIKIPEDIA_MODEL_ENCODING));
-                final HttpGet get = new HttpGet(url);
-                final HttpResponse response = httpclient.execute(get);
-                final StatusLine statusLine = response.getStatusLine();
-                if (statusLine.getStatusCode() == HttpStatus.SC_OK) {
-                    final ByteArrayOutputStream out = new ByteArrayOutputStream();
-                    response.getEntity().writeTo(out);
-                    out.close();
-                    responseString = out.toString(WIKIPEDIA_MODEL_ENCODING);
-                } else {
-                    // Closes the connection.
-                    response.getEntity().getContent().close();
-                    throw new IOException(statusLine.getReasonPhrase());
+    private void configureWebView() {
+        webView.setWebViewClient(new WebViewClient() {
+            @Override
+            public boolean shouldOverrideUrlLoading(WebView view, String url) {
+                if (Uri.parse(url).getHost().equals("www.wikipedia.org")) {
+                    return false;
                 }
-            } catch (ClientProtocolException e) {
-                onRetriveWikipediaUrlFailure();
-            } catch (IOException e) {
-                onRetriveWikipediaUrlFailure();
+                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(url));
+                startActivity(intent);
+                return true;
             }
-            return responseString;
-        }
-
-        @Override
-        protected void onPostExecute(final String result) {
-            super.onPostExecute(result);
-            try {
-                if (result.length() > 0) {
-                    final JSONObject jObject = new JSONObject(result);
-                    final JSONObject query = jObject.getJSONObject("query");
-                    final JSONArray search = query.getJSONArray("search");
-                    if (search.length() > 0) {
-                        // take the first result ...
-                        final JSONObject wikipediaArticle = (JSONObject) search
-                                .get(0);
-                        final String title = wikipediaArticle.getString("title");
-                        onRetriveWikipediaUrlSuccess(String.format(responseUrl,
-                                title.replace(" ", "_")));
-                    } else {
-                        onRetriveWikipediaUrlFailure();
-                    }
-                } else {
-                    onRetriveWikipediaUrlFailure();
-                }
-            } catch (JSONException e) {
-                onRetriveWikipediaUrlFailure();
-            }
-        }
+        });
+        WebSettings webSettings = webView.getSettings();
+        webSettings.setJavaScriptEnabled(true);
     }
 }
